requirements.txt
amqp==2.1.1
billiard==3.5.0.2
celery==4.0.0
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
kombu==4.0.0
MarkupSafe==0.23
numpy==1.11.1
Pillow==3.3.1
pytesseract==0.1.6
pytz==2016.7
redis==2.10.5
vine==1.1.3
Werkzeug==0.11.11
